Extract dividend and interest processing into cgt_calc/income.py
Third step of splitting main.py into phase modules. Moves the eight
dividend/interest methods from CapitalGainsCalculator into an
IncomeProcessor that takes the shared CalculatorState plus the
converters it reads; the calculator keeps a process_dividends
delegate for external callers. The dividend tax attribution memo
moves off CalculatorState onto the processor, where the derived
value belongs, dropping the SLF001 suppressions PR 2 left.
Four caplog filters follow the moved warnings to the cgt_calc.income
logger. Pure move, no behavior change.
